refactor: auditing 적용
